Redeploy VPN client when dirty
The problem arises when if_wg(4) is not loaded, so the VPN client
is created and started, but Wireguard does not start. The .done
file is not written so the jail is "dirty" (at least for this Makejail),
the problem is that said file is only verified when the jail is not
started, which is not correct in some cases since The jail can be
"dirty" even when it is started.