Fix $HOMEDIR/.config/littlejet/config.conf: No such file or directory
If the user configuration is not created, the error "No such file or directory"
is displayed for each run, although LittleJet works even with such an error
(but it's annoying, of course). The solution could be to just check before
loading, but I think this is a good opportunity to automatically create a
user config file if it doesn't exist, so I did it by putting some parameters
that can probably be suitable for most users, but not for some.